--- react-front-end/src/components/Component.js
@@ -1,1 +1,0 @@
-//This is just a sample file for the component folder to be created.
